+++ project/include/support.h
@@ -4,7 +4,7 @@
 #define PROJECT_INCLUDE_SUPPORT_H_
 
 #include <stddef.h>
-#include <utils.h>
+#include <matrix.h>
 
 Matrix* get_minor(const Matrix* matrix, size_t row, size_t col);
 int check_for_exist(const Matrix *matrix);
