+++ project/src/extra.c
@@ -35,24 +35,33 @@ int det(const Matrix* matrix, double* val) {
     return 0;
 }
 
-Matrix* get_minor(const Matrix* matrix, size_t row, size_t col) {
-    if (!matrix) {
-    return NULL;
+Matrix* adj(const Matrix* matrix) {
+    if (check_for_exist(matrix))
+        return NULL;
+    if ((matrix->num_rows != matrix->num_cols) || (matrix ->num_cols == 0)) {
+        return NULL;
     }
-    Matrix* minor = create_matrix(matrix -> num_rows - 1, matrix -> num_cols - 1);
-    if (minor -> value && matrix -> value) {
-        size_t n = matrix -> num_rows;
-        int pointer = 0;
-        for (size_t i = 0; i < n; i++) {
-            if (i != row) {
-                 for (size_t j = 0; j < n; j++) {
-                    if (j != col) {
-                        *(minor -> value + pointer) = matrix -> value[i * n + j];
-                        pointer++;
-                    }
-                }
+    size_t n = matrix->num_rows;
+    if (n < 1) {
+         puts("uncountable det");
+         return NULL;
+    }
+    Matrix* adject = create_matrix(n, n);
+    if (n == 1) {
+        adject -> value[0] = 1;
+        return adject;
+    }
+    double minor_det;
+    for (size_t i = 0; i < n; i++) {
+        for (size_t j = 0; j < n; j++) {
+            Matrix* minor = get_minor(matrix, j, i);
+            if ((minor != NULL) && !(det(minor, &minor_det))) {
+                int minor_sign = ((i + j) % 2)? -1: 1;
+                adject -> value[i * n + j] = minor_sign * minor_det;
             }
+        free_matrix(minor);
         }
     }
-    return minor;
+    return adject;
 }
+
