--- project/src/free_matrix.c
@@ -1,13 +1,0 @@
-
-#include <stdio.h>
-#include <stdlib.h>
-#include "matrix.h"
-
-void free_matrix(Matrix* matrix) {
-    if (matrix != NULL) {
-        if (matrix -> value != NULL) {
-            free(matrix -> value);
-        }
-    free(matrix);
-    }
-}
